サーバーTSに変更 #6
