+++ package.json
@@ -8,7 +8,7 @@
     "start": "ng serve",
     "deploy": "run-s build:client cpapp",
     "build:client": "ng build",
-    "build:clean":"ng build",
+    "build:clean": "ng build",
     "cpapp": "cp -Rf ./dist/* ../server/app",
     "test": "ng test",
     "pree2e": "webdriver-manager update --standalone false --gecko false",
@@ -34,25 +34,30 @@
     "ag-grid-angular": "^17.1.0",
     "angular-in-memory-web-api": "^0.6.1",
     "bootstrap": "^4.1.3",
+    "cookie-parser": "~1.4.3",
     "core-js": "^2.5.7",
+    "debug": "~2.6.9",
+    "express": "~4.16.0",
+    "helmet": "^3.15.0",
+    "http-errors": "~1.6.2",
     "jquery": "^3.3.1",
+    "morgan": "~1.9.0",
+    "mysql": "^2.16.0",
+    "nodemon": "^1.18.3",
     "popper.js": "^1.14.3",
+    "pug": "2.0.0-beta11",
     "rxjs": "^6.2.2",
     "rxjs-compat": "^6.0.0-rc.0",
     "ts-helpers": "^1.1.1",
-    "zone.js": "^0.8.26",
-    "cookie-parser": "~1.4.3",
-    "debug": "~2.6.9",
-    "express": "~4.16.0",
-    "http-errors": "~1.6.2",
-    "morgan": "~1.9.0",
-    "nodemon": "^1.18.3",
-    "pug": "2.0.0-beta11"
+    "winston": "^3.1.0",
+    "winston-daily-rotate-file": "^3.5.1",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.7.0",
     "@angular/cli": "^6.1.0",
     "@angular/compiler-cli": "^6.1.0",
+    "@types/express": "^4.16.0",
     "@types/jasmine": "2.8.6",
     "@types/node": "^9.6.23",
     "codelyzer": "~4.2.1",
