--- server/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "server",
-  "version": "0.0.0",
-  "private": true,
-  "scripts": {
-    "start": "nodemon ./bin/www"
-  },
-  "dependencies": {
-    "cookie-parser": "~1.4.3",
-    "debug": "~2.6.9",
-    "express": "~4.16.0",
-    "http-errors": "~1.6.2",
-    "morgan": "~1.9.0",
-    "nodemon": "^1.18.3",
-    "pug": "2.0.0-beta11"
-  },
-  "devDependencies": {}
-}
